moved class Article to article package
